requirements.txt
Flask==0.10.1
Flask-SocketIO==1.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pyserial==3.0
python-engineio==0.8.5
python-socketio==0.9.0
six==1.10.0
Werkzeug==0.11.3
wheel==0.26.0
